pyproject.toml
[tool.poetry]
name = "web-automation-lambdatest-pytest"
version = "0.1.0"
description = ""
authors = ["Jonathan Brull Schroeder <jonathans@avalith.net>"]

[tool.poetry.dependencies]
python = "^3.9"
pyleniumio = "^1.12.8"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"
flake8 = "^3.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
